requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt

flask==3.1.0  # Framework base
flask-login==0.6.3  # Manejo de sesiones de usuario
flask-wtf==1.2.2  # Formularios seguros
flask-sqlalchemy-lite==0.1.0  # Base de datos
sqlalchemy==2.0.37  # ORM para la base de datos
python-dotenv==1.0.1  # Variables de entorno
werkzeug==3.1.3  # Utilidades de Flask
wtforms==3.2.1  # Formularios
bcrypt==4.2.1  # Para encriptar contraseñas
flask-security==5.6.0  # Seguridad y autenticación
email-validator==2.2.0  # Validar correos en formularios
alembic==1.14.0  # Control de versiones de la base de datos